pyproject.toml
[tool.poetry]
name = "employees-sql-chatbot"
version = "0.0.1"
description = "An intelligent SQL query translation and data exploration platform"
authors = ["Nitin Jain <nitinbljain@outlook.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.11.5"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
streamlit = "^1.21.0"
sqlalchemy = "^1.4.48"
pandas = "^2.0.0"
langchain = "0.0.267"
openai = "^0.27.4"
pydantic = "1.10.7"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
psycopg2-binary = { version = "^2.9.5", optional = true }
azure-identity = "^1.24.0"
polars = "^1.33.0"

[tool.poetry.extras]
postgresql = ["psycopg2-binary"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
bandit = "^1.7.5"
safety = "^2.3.5"
locust = "^2.15.1"
httpx = "^0.24.0"  # For TestClient in FastAPI tests

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 120